+++ @xen-orchestra/xva/_writeDisk.mjs
@@ -1,22 +1,7 @@
 import { DiskLargerBlock, DiskSmallerBlock } from '@xen-orchestra/disk-transform'
-import { fromCallback } from 'promise-toolbox'
 import { formatBlockPath } from './_formatBlockPath.mjs'
-import { xxhash64 } from 'hash-wasm'
-
-export const XVA_DISK_CHUNK_LENGTH = 1024 * 1024
-async function addEntry(pack, name, buffer) {
-  await fromCallback.call(pack, pack.entry, { name }, buffer)
-}
-
-async function writeBlock(pack, data, name) {
-  if (data.length < XVA_DISK_CHUNK_LENGTH) {
-    data = Buffer.concat([data, Buffer.alloc(XVA_DISK_CHUNK_LENGTH - data.length, 0)])
-  }
-  await addEntry(pack, name, data)
-  // weirdly, ocaml and xxhash return the bytes in reverse order to each other
-  const hash = (await xxhash64(data)).toString('hex').toUpperCase()
-  await addEntry(pack, `${name}.xxhash`, Buffer.from(hash, 'utf8'))
-}
+import { writeChunk } from './_writeChunkInXva.mjs'
+import { XVA_DISK_CHUNK_LENGTH } from './_constants.mjs'
 
 export default async function addDisk(pack, disk, basePath) {
   let diskXvaChunk
@@ -42,7 +27,7 @@ export default async function addDisk(pack, disk, basePath) {
       mustWriteBlock = mustWriteBlock || !block.data.equals(empty)
     }
     if (mustWriteBlock) {
-      await writeBlock(pack, block.data, formatBlockPath(basePath, counter))
+      await writeChunk(pack, block.data, formatBlockPath(basePath, counter))
     }
   }
 }
